lock added

+++ kernel/defs.h
@@ -63,7 +63,7 @@ void            ramdiskrw(struct buf*);
 void*           kalloc(void);
 void            kfree(void *);
 void            kinit(void);
-int             amountOfFreeMemory(void);
+uint64             amountOfFreeMemory(void);
 
 // log.c
 void            initlog(int, struct superblock*);
